ContactListViewModel: improve performance of fetching contacts (#1)
For users who have a lot of contacts, the background process for
fetching and analyzing the full list of contacts is too resource
intensive. The operating system has strict limits on how much work
a background task can do. When a process is too demanding for too long,
the OS can terminate it prematurely to protect system performance and
battery life. Because the task is terminated before it can finish
analyzing all your contacts and report back, the main application never
receives the list of contacts that need formatting. It's left with an
empty list, and therefore incorrectly shows the "All contact phone
numbers are formatted correctly" message.

Instead of trying to process all contacts in one single, monolithic
background task, this change updates the code to process them in
a stream, allowing the app to work through user contacts incrementally
in smaller, more manageable chunks.